html and css templates
So here's the blank html template with the bottom navigation bar, and the basic css file to go with it. The css is mostly just copied stuff from old project since we haven't agreed on any standard font sizes colors and such, so that can be changed if needed. I got the Icons from https://icons8.com/. They can be changed whenever. I also changed the favorites button to matchmaking as was planned. Also copied fonts folder from old project